+++ connectors/core/cloud-connector/src/test/java/org/platformlambda/cloud/ConnectorTest.java
@@ -19,7 +19,6 @@
 package org.platformlambda.cloud;
 
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.platformlambda.cloud.reporter.PresenceConnector;
 import org.platformlambda.core.models.EventEnvelope;
@@ -167,19 +166,6 @@ public class ConnectorTest extends TestBase {
         queryResult = bench.poll(10, TimeUnit.SECONDS);
         assert queryResult != null;
         assertEquals(true, queryResult.getBody());
-        headers.put("X-App-Instance", Platform.getInstance().getOrigin());
-        response = httpPost("http://127.0.0.1:"+port, "/suspend/now", headers, new HashMap<>());
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(200, result.get("status"));
-        assertEquals("suspend", result.get("type"));
-        assertEquals("/suspend/now", result.get("path"));
-        response = httpPost("http://127.0.0.1:"+port, "/resume/now", headers, new HashMap<>());
-        assertInstanceOf(Map.class, response.getBody());
-        result = (Map<String, Object>) response.getBody();
-        assertEquals(200, result.get("status"));
-        assertEquals("resume", result.get("type"));
-        assertEquals("/resume/now", result.get("path"));
         po.send(ServiceDiscovery.SERVICE_REGISTRY, new Kv("type", "leave"), new Kv("origin", origin));
     }