+++ system/platform-core/src/test/java/org/platformlambda/automation/AdminEndpointTest.java
@@ -68,33 +68,6 @@ public class AdminEndpointTest extends TestBase {
 
     @SuppressWarnings("unchecked")
     @Test
-    public void nonExistRemoteInfoEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpGet(localHost, "/info", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void protectInfoEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        EventEnvelope response = httpGet("http://localhost:"+port, "/info", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("Resource not found", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
     public void libEndpointTest() throws IOException, InterruptedException {
         EventEnvelope response = httpGet(localHost, "/info/lib", null);
         assert response != null;
@@ -107,20 +80,6 @@ public class AdminEndpointTest extends TestBase {
 
     @SuppressWarnings("unchecked")
     @Test
-    public void remoteLibEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpGet(localHost, "/info/lib", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
     public void routeEndpointTest() throws IOException, InterruptedException {
         EventEnvelope response = httpGet(localHost, "/info/routes", null);
         assert response != null;
@@ -129,20 +88,6 @@ public class AdminEndpointTest extends TestBase {
         assertInstanceOf(Map.class, result.get("routing"));
         Map<String, Object> routing = (Map<String, Object>) result.get("routing");
         assertEquals(new HashMap<>(), routing.get("routes"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void remoteRouteEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpGet(localHost, "/info/routes", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
     }
 
     @SuppressWarnings("unchecked")
@@ -194,34 +139,6 @@ public class AdminEndpointTest extends TestBase {
 
     @SuppressWarnings("unchecked")
     @Test
-    public void remoteHealthEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpGet(localHost, "/health", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void remoteLivenessEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpGet(localHost, "/livenessprobe", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
     public void envEndpointTest() throws IOException, InterruptedException {
         EventEnvelope response = httpGet(localHost, "/env", null);
         assert response != null;
@@ -232,142 +149,6 @@ public class AdminEndpointTest extends TestBase {
         assertInstanceOf(Map.class, multi.getElement("env"));
         assertInstanceOf(List.class, multi.getElement("routing.private"));
         assertInstanceOf(List.class, multi.getElement("routing.public"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void remoteEnvEndpointTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpGet(localHost, "/livenessprobe", headers);
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @Test
-    public void shutdownUsingGetWillFail() throws IOException, InterruptedException {
-        EventEnvelope response = httpGet(localHost, "/shutdown/now", null);
-        assert response != null;
-        assertEquals(404, response.getStatus());
-    }
-
-    @Test
-    public void suspendUsingGetWillFail() throws IOException, InterruptedException {
-        EventEnvelope response = httpGet(localHost, "/suspend/now", null);
-        assert response != null;
-        assertEquals(404, response.getStatus());
-    }
-
-    @Test
-    public void resumeUsingGetWillFail() throws IOException, InterruptedException {
-        EventEnvelope response = httpGet(localHost, "/resume/now", null);
-        assert response != null;
-        assertEquals(404, response.getStatus());
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void shutdownWithoutAppInstanceWillFail() throws IOException, InterruptedException {
-        EventEnvelope response = httpPost(localHost, "/shutdown", null, new HashMap<>());
-        assert response != null;
-        assertEquals(400, response.getStatus());
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals("Missing X-App-Instance in request header", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void shutdownWithIncorrectAppInstanceWillFail() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpPost(localHost, "/shutdown", headers, new HashMap<>());
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void suspendWithoutAppInstanceWillFail() throws IOException, InterruptedException {
-        EventEnvelope response = httpPost(localHost, "/suspend/now", null, new HashMap<>());
-        assert response != null;
-        assertEquals(400, response.getStatus());
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals("Missing X-App-Instance in request header", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void suspendWithIncorrectAppInstanceWillFail() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpPost(localHost, "/suspend/now", headers, new HashMap<>());
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void resumeWithoutAppInstanceWillFail() throws IOException, InterruptedException {
-        EventEnvelope response = httpPost(localHost, "/resume/now", null, new HashMap<>());
-        assert response != null;
-        assertEquals(400, response.getStatus());
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals("Missing X-App-Instance in request header", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void resumeWithIncorrectAppInstanceWillFail() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", "does-not-exist");
-        EventEnvelope response = httpPost(localHost, "/resume/now", headers, new HashMap<>());
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals(404, response.getStatus());
-        assertEquals(404, result.get("status"));
-        assertEquals("does-not-exist is not reachable", result.get("message"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void suspendTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", Platform.getInstance().getOrigin());
-        EventEnvelope response = httpPost(localHost, "/suspend/now", headers, new HashMap<>());
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals("suspend", result.get("type"));
-        assertEquals(200, result.get("status"));
-    }
-
-    @SuppressWarnings("unchecked")
-    @Test
-    public void resumeTest() throws IOException, InterruptedException {
-        Map<String, String> headers = new HashMap<>();
-        headers.put("x-app-instance", Platform.getInstance().getOrigin());
-        EventEnvelope response = httpPost(localHost, "/resume/now", headers, new HashMap<>());
-        assert response != null;
-        assertInstanceOf(Map.class, response.getBody());
-        Map<String, Object> result = (Map<String, Object>) response.getBody();
-        assertEquals("resume", result.get("type"));
-        assertEquals(200, result.get("status"));
     }
 
     @Test
@@ -431,5 +212,4 @@ public class AdminEndpointTest extends TestBase {
         assertEquals(404, result.get("status"));
         assertEquals("Resource not found", result.get("message"));
     }
-
 }