+++ CHANGELOG.md
@@ -8,6 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.2.13, 2/13/2025
+
+### Added
+
+Actuator REST endpoints are now configurable in rest.yaml
+
+### Removed
+
+The feature to shutdown, suspend and resume of an application instance is retired
+
+### Changed
+
+Update acturator services to serve REST requests directly
+
+---
 ## Version 4.2.12, 2/12/2025
 
 ### Added